Add files via upload
In search of a meaningful data-based capstone project, I reached out to several developer friends to see if they had any suggestions. One such friend relayed his excitement over a pirate themed digital card game he was developing which he was in the process of play testing. \He finished the backend work and developed a simple AI to run through thousands of simulations to see if there was anything that was disbalanced or overpowered.
He got the data in JSON format and tasked me with several questions that could help him make gameplay with users more fun and run more smoothly. In the R and Python files I've left comments to better understand the process of the analysis. A big goal was to allow for my friend to reuse this code with his different dataframes after different simulations.